[Import Data File Command] Use CSV reader to read csv files for correct parsing issues of characters like newlines and tabs (#747)
- Use csv reader to read csv files to avoid parsing issues with characters like newlines and tabs
- Added a flag --truncate-splits set to false, if required, for debugging purposes
- Added automation tests for import file from csv format
- Introduced a new env variable 'CSV_NO_NEWLINE' to support csv formats with any given escape/quote characters but without any newline characters in data file

+++ yb-voyager/cmd/importData.go
@@ -75,6 +75,7 @@ var copyTableFromCommands = make(map[string]string)
 var loadBalancerUsed bool           // specifies whether load balancer is used in front of yb servers
 var enableUpsert bool               // upsert instead of insert for import data
 var disableTransactionalWrites bool // to disable transactional writes for copy command
+var truncateSplits bool             // to truncate *.D splits after import
 
 const (
 	LB_WARN_MSG = "Warning: Based on internal anaylsis, --target-db-host is identified as a load balancer IP which will be used to create connections for data import.\n" +
@@ -994,9 +995,12 @@ func markTaskDone(task *SplitFileImportTask) {
 	if err != nil {
 		utils.ErrExit("rename %q => %q: %s", inProgressFilePath, doneFilePath, err)
 	}
-	err = os.Truncate(doneFilePath, 0)
-	if err != nil {
-		log.Warnf("truncate file %q: %s", doneFilePath, err)
+
+	if truncateSplits {
+		err = os.Truncate(doneFilePath, 0)
+		if err != nil {
+			log.Warnf("truncate file %q: %s", doneFilePath, err)
+		}
 	}
 }
